Completed UsePopcorn and added custom hooks
