+++ src/useKey.jsx
@@ -1,0 +1,19 @@
+import { useEffect } from "react";
+
+export function useKey(key, action) {
+  useEffect(
+    function () {
+      function callback(e) {
+        if (e.code.toLowerCase() === key.toLowerCase()) {
+          console.log("CLOSE");
+          action();
+        }
+      }
+      document.addEventListener("keydown", callback);
+      return function () {
+        document.removeEventListener("keydown", callback);
+      };
+    },
+    [action, key]
+  );
+}
